+++ pkg/handler/admin_tasks.go
@@ -3,10 +3,12 @@ package handler
 import (
 	"net/http"
 
+	"github.com/content-services/content-sources-backend/pkg/api"
 	"github.com/content-services/content-sources-backend/pkg/dao"
 	ce "github.com/content-services/content-sources-backend/pkg/errors"
 	"github.com/content-services/content-sources-backend/pkg/rbac"
 	"github.com/labstack/echo/v4"
+	"github.com/rs/zerolog/log"
 )
 
 type AdminTaskHandler struct {
@@ -57,4 +59,23 @@ func (adminTaskHandler *AdminTaskHandler) fetch(c echo.Context) error {
 		return ce.NewErrorResponse(ce.HttpCodeForDaoError(err), "Error fetching task", err.Error())
 	}
 	return c.JSON(http.StatusOK, response)
+}
+
+func ParseAdminTaskFilters(c echo.Context) api.AdminTaskFilterData {
+	filterData := api.AdminTaskFilterData{
+		AccountId: DefaultAccountId,
+		OrgId:     DefaultOrgId,
+		Status:    DefaultStatus,
+	}
+	err := echo.QueryParamsBinder(c).
+		String("account_id", &filterData.AccountId).
+		String("org_id", &filterData.OrgId).
+		String("status", &filterData.Status).
+		BindError()
+
+	if err != nil {
+		log.Error().Err(err).Msg("Error parsing filters")
+	}
+
+	return filterData
 }
